playground/pyproject.toml
[project]
name = "playground"
version = "0.1.0"
description = "お遊び環境。API実験、プロンプト試行、ライブラリ検証など自由に使う"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.66.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
]
